Whitelist the (in)famous HORI mini controller...
and potentially add the ability to whitelist others.

This unfortunately does not, however, enables the user to use a HORI mini as a
passthrough device. For some unknown reason, the HORI mini sends the wrong
signature for the challenge data after the first authentication. I suspect
that this is due to a bug somewhere in the HORI mini's implementation of DS4
authentication. It is unknown why this controller works well when it
connects to the PS4 directly. (perhaps a workaround on the PS4/Sony side exists
that detects this controller and relaxes the challenge?)